sp1Decor, chardirection, fix collide leftright

+++ mvtPatterns.h
@@ -2,6 +2,6 @@
 #define GLITCH_MVT_PATTERNS_H_
 
 //hyberbole y value for x*x + y = 0 (x between -11 & 12)
-static const int8_t JUMP_VY_PATTERN[10] = {-25,-16,-9,-4,-1,1,4,9,16,25};
+static const float JUMP_VY_PATTERN[10] = {-25.0, -16.0, -9.0, -4.0, -1.0, 1.0, 4.0, 9.0, 16.0, 25.0};
 
 #endif
